requirements.txt
transformers
lm_eval
datasets
nvidia-modelopt
